Update README.md
The list consists of my top 5 favorite movies of all time. I am a big fan of the cyberpunk genre and Akira is #1 on my list because it is the progenitor of the cyberpunk genre. Without this movie, the cyberpunk genre may or may have not existed in this timeline.